* Force load of `BouncyCastleProvider` to ensure `MD4` implementation  exists * Numerous code cleanups, I hope I didn't break anything * Use a `LOGGER` per class instead of using `JISystem.getLogger` * Deprecate `JISystem.setInBuiltLogHandler` * Disable execution of "tests" files when compiling

+++ j-interopdeps/src/main/java/rpc/pdu/AlterContextPdu.java
@@ -73,7 +73,7 @@ public class AlterContextPdu extends ConnectionOrientedPdu {
     protected void readBody(NetworkDataRepresentation ndr) {
         setMaxTransmitFragment(ndr.readUnsignedShort());
         setMaxReceiveFragment(ndr.readUnsignedShort());
-        setAssociationGroupId((int) ndr.readUnsignedLong());
+        setAssociationGroupId(ndr.readUnsignedLong());
         int count = ndr.readUnsignedSmall();
         PresentationContext[] contextList = new PresentationContext[count];
         for (int i = 0; i < count; i++) {
